+++ crates/spark-model/src/layers/qwen3_ssm/qkvz_projection.rs
@@ -1,0 +1,179 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+//! QKVZ projection (+ deinterleave) step of
+//! `Qwen3SsmLayer::decode_batched_inner`. Extracted from
+//! `trait_decode_batched.rs` to keep the parent file under 500 LoC.
+
+use super::*;
+
+impl Qwen3SsmLayer {
+    /// For `sequential_qkvz` (Qwen3.5): writes directly to `deinterleaved`.
+    /// For interleaved (80B): writes to `ssm_qkvz()`, then deinterleaves per
+    /// token into `deinterleaved`.
+    #[allow(clippy::too_many_arguments)]
+    pub(super) fn qkvz_projection_batched(
+        &self,
+        ctx: &ForwardContext,
+        normed: DevicePtr,
+        deinterleaved: DevicePtr,
+        num_tokens: usize,
+        k: u32,
+        h: usize,
+        qkvz_size: usize,
+        nk: usize,
+        kd: usize,
+        vpg: usize,
+        vd: usize,
+        stream: u64,
+    ) -> Result<()> {
+        let bf16 = 2usize;
+        let proj_dst = if self.sequential_qkvz {
+            deinterleaved
+        } else {
+            ctx.buffers.ssm_qkvz()
+        };
+        if num_tokens == 3 {
+            if let Some(ref nvfp4) = self.qkvz_nvfp4 {
+                ops::w4a16_gemv_batch3(
+                    ctx.gpu,
+                    self.w4a16_gemv_batch3_k,
+                    normed,
+                    nvfp4,
+                    proj_dst,
+                    qkvz_size as u32,
+                    h as u32,
+                    stream,
+                )?;
+            } else {
+                for t in 0..3u32 {
+                    ops::dense_gemv(
+                        ctx.gpu,
+                        self.dense_gemv_k,
+                        normed.offset(t as usize * h * bf16),
+                        &self.ssm.in_proj_qkvz,
+                        proj_dst.offset(t as usize * qkvz_size * bf16),
+                        qkvz_size as u32,
+                        h as u32,
+                        stream,
+                    )?;
+                }
+            }
+        } else if num_tokens == 2 {
+            if let Some(ref nvfp4) = self.qkvz_nvfp4 {
+                ops::w4a16_gemv_batch2(
+                    ctx.gpu,
+                    self.w4a16_gemv_batch2_k,
+                    normed,
+                    nvfp4,
+                    proj_dst,
+                    qkvz_size as u32,
+                    h as u32,
+                    stream,
+                )?;
+            } else {
+                ops::dense_gemv(
+                    ctx.gpu,
+                    self.dense_gemv_k,
+                    normed,
+                    &self.ssm.in_proj_qkvz,
+                    proj_dst,
+                    qkvz_size as u32,
+                    h as u32,
+                    stream,
+                )?;
+                ops::dense_gemv(
+                    ctx.gpu,
+                    self.dense_gemv_k,
+                    normed.offset(h * bf16),
+                    &self.ssm.in_proj_qkvz,
+                    proj_dst.offset(qkvz_size * bf16),
+                    qkvz_size as u32,
+                    h as u32,
+                    stream,
+                )?;
+            }
+        } else if let Some(fp8) = self.qkvz_fp8 {
+            ops::fp8_gemm_n128(
+                ctx.gpu,
+                self.fp8_gemm_k,
+                normed,
+                fp8,
+                proj_dst,
+                k,
+                qkvz_size as u32,
+                h as u32,
+                stream,
+            )?;
+        } else if let Some(ref nvfp4_t) = self.qkvz_nvfp4_t {
+            // m128 halves B re-reads for large M (prefill); falls back to m64 for M≤128.
+            if k > 128 {
+                ops::w4a16_gemm_n128_m128(
+                    ctx.gpu,
+                    self.w4a16_gemm_t_m128_k,
+                    normed,
+                    nvfp4_t,
+                    proj_dst,
+                    k,
+                    qkvz_size as u32,
+                    h as u32,
+                    stream,
+                )?;
+            } else {
+                ops::w4a16_gemm_n128(
+                    ctx.gpu,
+                    self.w4a16_gemm_t_k,
+                    normed,
+                    nvfp4_t,
+                    proj_dst,
+                    k,
+                    qkvz_size as u32,
+                    h as u32,
+                    stream,
+                )?;
+            }
+        } else if let Some(ref nvfp4) = self.qkvz_nvfp4 {
+            ops::w4a16_gemm(
+                ctx.gpu,
+                self.w4a16_gemm_k,
+                normed,
+                nvfp4,
+                proj_dst,
+                k,
+                qkvz_size as u32,
+                h as u32,
+                stream,
+            )?;
+        } else {
+            ops::dense_gemm(
+                ctx.gpu,
+                self.dense_gemm_k,
+                normed,
+                &self.ssm.in_proj_qkvz,
+                proj_dst,
+                k,
+                qkvz_size as u32,
+                h as u32,
+                stream,
+            )?;
+        }
+        if !self.sequential_qkvz {
+            for t in 0..(num_tokens as u32) {
+                let src = proj_dst.offset(t as usize * qkvz_size * bf16);
+                let dst = deinterleaved.offset(t as usize * qkvz_size * bf16);
+                ops::deinterleave_qkvz(
+                    ctx.gpu,
+                    self.deinterleave_k,
+                    src,
+                    dst,
+                    1,
+                    nk as u32,
+                    kd as u32,
+                    vpg as u32,
+                    vd as u32,
+                    stream,
+                )?;
+            }
+        }
+        Ok(())
+    }
+}
